+++ packages/dso-toolkit/src/components/ozon-content/ozon-content.content.ts
@@ -492,6 +492,9 @@ export const content: OzonContentStory[] = [
             uitlijning="start"
             alt="Afbeelding 1"
             schaal="50"
+            hoogte="1124"
+            breedte="720"
+            dpi="150"
           />
           <Bijschrift locatie="onder">Bijschrift van het figuur.</Bijschrift>
           <Bron>Bron waaruit het figuur is overgenomen</Bron>
@@ -595,10 +598,12 @@ export const content: OzonContentStory[] = [
                   wId="gm1979_2__chp_13__subsec_13.7__art_13.72__table_o_1__img_o_1"
                 >
                   <Illustratie
-                    naam="https://nep-knooppunt-test.viewer.dso.kadaster.nl/publiek/omgevingsdocumenten/api/presenteren/v7/afbeeldingen/gm1979/_akn_nl_act_gm1979_2021_omgevingsplandelfzijl_nld_2/image1.png"
-                    breedte="7813"
-                    hoogte="4228"
-                    uitlijning="start"
+                    breedte='7813'
+                    dpi='72'
+                    formaat='image/png'
+                    hoogte='4228'
+                    naam='https://kta.document-viewer-api.dso.kadaster.nl/ozon/presenteren/v7/ontwerpafbeeldingen/gm1979/_akn_nl_bill_gm1979_2021_omgevingsplandelfzijl_ZeventiendeOntwerp/image1.png'
+                    uitlijning='start'
                   />
                 </Figuur>
               </entry>
@@ -1715,13 +1720,13 @@ export const content: OzonContentStory[] = [
           <Inhoud>
             <Figuur eId='chp_17__subchp_17.1__subsec_17.1.1__subsec_17.1.1.1__subsec_17.1.1.1.1__art_17.1__para_4__img_o_1_inst2' wId='gm0344_499f564bbf174298975d7d7ece6b61e2__chp_17__subchp_17.1__subsec_17.1.1__subsec_17.1.1.1__subsec_17.1.1.1.1__art_17.1__para_4__img_o_1_inst2' wijzigactie='verwijder'>
                 <Titel>Klimaatambities</Titel>
-                <Illustratie naam='images/klimaatambities_agv_infographic_002_1.jpg' breedte='366' hoogte='206'/>
+                <Illustratie naam='images/klimaatambities_agv_infographic_002_1.jpg' breedte='366' hoogte='206' dpi="72"/>
                 <Bijschrift locatie='onder'>test</Bijschrift>
                 <Bron>www.google.nl</Bron>
             </Figuur>
             <Figuur eId='chp_18__subchp_18.1__subsec_18.1.1__subsec_18.1.1.1__subsec_18.1.1.1.1__art_18.1__para_1__img_o_1' wId='gm0344_499f564bbf174298975d7d7ece6b61e2__chp_18__subchp_18.1__subsec_18.1.1__subsec_18.1.1.1__subsec_18.1.1.1.1__art_18.1__para_1__img_o_1' wijzigactie='voegtoe'>
                 <Titel>Klimaatambities</Titel>
-                <Illustratie naam='images/klimaatambities_agv_infographic_002_1.jpg' breedte='366' hoogte='206'/>
+                <Illustratie naam='images/klimaatambities_agv_infographic_002_1.jpg' breedte='366' hoogte='206' dpi="72"/>
                 <Bijschrift locatie='onder'>test</Bijschrift>
                 <Bron>www.google.nl</Bron>
             </Figuur>